# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/investment-requests": {
      "get": {
        "operationId": "InvestmentRequestController_getAllinvestmentRequests",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestmentRequestGetBusinessOwnerResponse"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "InvestmentRequestController_createInvestmentRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestmentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/string"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}": {
      "get": {
        "operationId": "InvestmentRequestController_getInvestmentRequest",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentRequestGetDetailedResponse"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "InvestmentRequestController_updateInvestmentRequest",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvestmentRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "InvestmentRequestController_deleteInvestmentRequest",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/rating-data": {
      "get": {
        "operationId": "InvestmentRequestController_getCreditRatingDataResponse",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreditRatingDataResponse"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/loan": {
      "get": {
        "operationId": "InvestmentRequestController_getLastLoanForBusinessOwner",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLoanBusinessOwnerResponse"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/admin/business-owner-loan/{businessOwnerId}": {
      "get": {
        "operationId": "InvestmentRequestController_getBusinessOwnersInstallmentsForAdmin",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "businessOwnerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLoanBusinessOwnerResponse"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/installments": {
      "get": {
        "operationId": "InvestmentRequestController_getLastLoanForInvestor",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLoanInvestorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/admin/investor-loan/{investorId}": {
      "get": {
        "operationId": "InvestmentRequestController_getInvestorInstallmentsForAdmin",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "investorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLoanInvestorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/accept": {
      "patch": {
        "operationId": "InvestmentRequestController_acceptLoan",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/string"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/decline": {
      "patch": {
        "operationId": "InvestmentRequestController_declineLoan",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/string"
                }
              }
            }
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/investment-requests/{investmentRequestId}/extend": {
      "patch": {
        "operationId": "InvestmentRequestController_extendRaisingTime",
        "parameters": [
          {
            "name": "investmentRequestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "investments requests"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin-data/global-settings": {
      "get": {
        "operationId": "GlobalSettingsController_getGlobalSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetGlobalSettingsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GlobalSettingsController_updateGlobalSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGlobalSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin-data/global-settings/ratings": {
      "get": {
        "operationId": "GlobalSettingsController_getRatings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RangesResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "GlobalSettingsController_updateRanges",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRatingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/admin-data/global-settings/payment-data": {
      "get": {
        "operationId": "GlobalSettingsController_getAdminPaymentData",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin panel data"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Asabik API Docs",
    "description": "All endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvestmentRequestGetBusinessOwnerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "purposeOfTheLoan": {
            "type": "string",
            "enum": [
              "Working capital",
              "Equipment Purchase",
              "Inventory Purchase",
              "Expand Business",
              "Finance Debt Obligation",
              "Marketing or Design",
              "Emergency",
              "Other"
            ]
          },
          "loanPurpose": {
            "type": "string"
          },
          "helpIncreaseProfit": {
            "type": "string"
          },
          "profitIncrease": {
            "type": "number"
          },
          "returnTerm": {
            "type": "number"
          },
          "requiredCapital": {
            "type": "number"
          },
          "netReturn": {
            "type": "number"
          },
          "netReturnToShare": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "Request is beeing processed",
              "Rejected - less than $2000 raised",
              "Application cancelled",
              "Application closed",
              "Funds has been transfered",
              "Application open",
              "Application awaits for manual processing",
              "Manual processing",
              "Open to invest",
              "Raising funds",
              "Closed",
              "Funds to be deposited"
            ]
          },
          "percentageRaised": {
            "type": "number"
          },
          "timeLeft": {
            "type": "string"
          },
          "isEligible": {
            "type": "boolean"
          },
          "canEdit": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "purposeOfTheLoan",
          "loanPurpose",
          "helpIncreaseProfit",
          "profitIncrease",
          "returnTerm",
          "requiredCapital",
          "netReturn",
          "netReturnToShare",
          "status",
          "percentageRaised",
          "timeLeft",
          "isEligible",
          "canEdit"
        ]
      },
...
      "CreateInvestmentRequestDto": {
        "type": "object",
        "properties": {
          "purposeOfTheLoan": {
            "type": "string",
            "enum": [
              "Working capital",
              "Equipment Purchase",
              "Inventory Purchase",
              "Expand Business",
              "Finance Debt Obligation",
              "Marketing or Design",
              "Emergency",
              "Other"
            ]
          },
          "requiredCapital": {
            "type": "number"
          },
          "returnTerm": {
            "type": "number"
          },
          "netReturn": {
            "type": "number"
          },
          "netReturnToShare": {
            "type": "number"
          },
          "loanPurpose": {
            "type": "string"
          },
          "helpIncreaseProfit": {
            "type": "string"
          },
          "profitIncrease": {
            "type": "number"
          }
        },
        "required": [
          "purposeOfTheLoan",
          "requiredCapital",
          "returnTerm",
          "netReturn",
          "netReturnToShare",
          "loanPurpose",
          "helpIncreaseProfit",
          "profitIncrease"
        ]
      },
...
      "InvestmentRequestGetDetailedResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "psr": {
            "type": "string"
          },
          "projectedNetReturn": {
            "type": "string"
          },
          "nextInstallmentAmount": {
            "type": "number"
          },
          "nextInstallmentDate": {
            "type": "string"
          },
          "nextPlaidVerifDate": {
            "type": "string"
          },
          "finalPaymentAmount": {
            "type": "number"
          },
          "finalPaymentDate": {
            "type": "string"
          },
          "dti": {
            "type": "string"
          },
          "rating": {
            "type": "string",
            "enum": [
              "Aaa",
              "Aa1",
              "Aa2",
              "Aa3",
              "A1",
              "A2",
              "A3",
              "Baa1",
              "Baa2",
              "Baa3",
              "Junk"
            ]
          },
          "approvedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "psr",
          "projectedNetReturn",
          "nextInstallmentAmount",
          "nextInstallmentDate",
          "nextPlaidVerifDate",
          "finalPaymentAmount",
          "finalPaymentDate",
          "dti",
          "rating",
          "approvedAt"
        ]
      },
      "CreditRatingDataResponse": {
        "type": "object",
        "properties": {
          "psr": {
            "type": "string"
          },
          "r": {
            "type": "number"
          },
          "m": {
            "type": "number"
          },
          "a": {
            "type": "number"
          },
          "returnTerm": {
            "type": "number"
          },
          "requestedAmount": {
            "type": "number"
          },
          "netReturn": {
            "type": "number"
          },
          "netReturnToShare": {
            "type": "number"
          },
          "totalNoMonth": {
            "type": "number"
          },
          "outflowExceed": {
            "type": "number"
          },
          "negativeBalance": {
            "type": "number"
          },
          "noEarning": {
            "type": "number"
          },
          "vInflow": {
            "type": "number"
          },
          "vTotal": {
            "type": "number"
          },
          "averageE": {
            "type": "number"
          },
          "averageL": {
            "type": "number"
          },
          "o": {
            "type": "number"
          },
          "i": {
            "type": "number"
          },
          "d1": {
            "type": "number"
          },
          "d2": {
            "type": "number"
          },
          "d0": {
            "type": "number"
          },
          "pod": {
            "type": "number"
          },
          "maxLoanCalc": {
            "type": "number"
          },
          "rating": {
            "type": "string",
            "enum": [
              "Aaa",
              "Aa1",
              "Aa2",
              "Aa3",
              "A1",
              "A2",
              "A3",
              "Baa1",
              "Baa2",
              "Baa3",
              "Junk"
            ]
          },
          "low": {
            "type": "number"
          },
          "high": {
            "type": "number"
          }
        },
        "required": [
          "psr",
          "r",
          "m",
          "a",
          "returnTerm",
          "requestedAmount",
          "netReturn",
          "netReturnToShare",
          "totalNoMonth",
          "outflowExceed",
          "negativeBalance",
          "noEarning",
          "vInflow",
          "vTotal",
          "averageE",
          "averageL",
          "o",
          "i",
          "d1",
          "d2",
          "d0",
          "pod",
          "maxLoanCalc",
          "rating",
          "low",
          "high"
        ]
      },
...
      "GetLoanBusinessOwnerResponse": {
        "type": "object",
        "properties": {
          "declaredMotnhlyProfit": {
            "type": "number"
          },
          "declaredProfitToShare": {
            "type": "number"
          },
          "psr": {
            "type": "string"
          },
          "installments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstallmentBusinessOwnerResponse"
            }
          }
        },
        "required": [
          "declaredMotnhlyProfit",
          "declaredProfitToShare",
          "psr",
          "installments"
        ]
      },
      "GetLoanInvestorResponse": {
        "type": "object",
        "properties": {
          "declaredProfitToShare": {
            "type": "number"
          },
          "raisedFunds": {
            "type": "number"
          },
          "invested": {
            "type": "number"
          },
          "investedPercentage": {
            "type": "string"
          },
          "installments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "declaredProfitToShare",
          "raisedFunds",
          "invested",
          "investedPercentage",
          "installments"
        ]
      },
      "UpdateInvestmentRequestDto": {
        "type": "object",
        "properties": {
          "loanPurpose": {
            "type": "string"
          },
          "helpIncreaseProfit": {
            "type": "string"
          },
          "netReturn": {
            "type": "number"
          },
          "netReturnToShare": {
            "type": "number"
          }
        },
        "required": [
          "loanPurpose",
          "helpIncreaseProfit",
          "netReturn",
          "netReturnToShare"
        ]
      },
      "string": {
        "type": "object",
        "properties": {}
      },
      "GetGlobalSettingsResponse": {
        "type": "object",
        "properties": {
          "maxReturnTermWOManualProcessing": {
            "type": "number"
          },
          "r": {
            "type": "number"
          },
          "m": {
            "type": "number"
          },
          "a": {
            "type": "number"
          },
          "raisingTimeLimit": {
            "type": "number"
          },
          "applicationFee": {
            "type": "number"
          },
          "minLoan": {
            "type": "number"
          },
          "maxLoan": {
            "type": "number"
          },
          "invalidTransactionPenalty": {
            "type": "number"
          },
          "plaidTokenPenalty": {
            "type": "number"
          }
        },
        "required": [
          "maxReturnTermWOManualProcessing",
          "r",
          "m",
          "a",
          "raisingTimeLimit",
          "applicationFee",
          "minLoan",
          "maxLoan",
          "invalidTransactionPenalty",
          "plaidTokenPenalty"
        ]
      },
...
      "UpdateGlobalSettingsDto": {
        "type": "object",
        "properties": {
          "maxReturnTermWOManualProcessing": {
            "type": "number"
          },
          "r": {
            "type": "number"
          },
          "m": {
            "type": "number"
          },
          "a": {
            "type": "number"
          },
          "raisingTimeLimit": {
            "type": "number"
          },
          "applicationFee": {
            "type": "number"
          },
          "minLoan": {
            "type": "number"
          },
          "maxLoan": {
            "type": "number"
          },
          "invalidTransactionPenalty": {
            "type": "number"
          },
          "plaidTokenPenalty": {
            "type": "number"
          }
        },
        "required": [
          "maxReturnTermWOManualProcessing",
          "r",
          "m",
          "a",
          "raisingTimeLimit",
          "applicationFee",
          "minLoan",
          "maxLoan",
          "invalidTransactionPenalty",
          "plaidTokenPenalty"
        ]
      },
...
      "UpdateRatingDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number"
          },
          "rangesRowUpdate": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RangesRowUpdateDto"
            }
          }
        },
        "required": [
          "year",
          "rangesRowUpdate"
        ]
      },
...
    }
  }
}